requirements.txt
Django>=1.10.2
mysqlclient>=1.3.9
django-debug-toolbar>=1.6
django-mailer>=1.2.2
django-allauth>=0.27.0
pytz
django-cleanup>=0.4.2
django-uuid-upload-path
django-activity-stream>=0.6.3